vanity refactor
