+++ frontend/pkg/frontend/operations.go
@@ -21,12 +21,178 @@ import (
 	"path"
 	"strings"
 
+	"github.com/Azure/ARO-HCP/internal/database"
+	"github.com/Azure/ARO-HCP/internal/ocm"
+	"github.com/Azure/ARO-HCP/internal/utils/armhelpers"
 	azcorearm "github.com/Azure/azure-sdk-for-go/sdk/azcore/arm"
+	cmv1 "github.com/openshift-online/ocm-sdk-go/clustersmgmt/v1"
 
 	"github.com/Azure/ARO-HCP/internal/api"
 	"github.com/Azure/ARO-HCP/internal/api/arm"
 	"github.com/Azure/ARO-HCP/internal/utils"
 )
+
+func (f *Frontend) OperationStatus(writer http.ResponseWriter, request *http.Request) error {
+	ctx := request.Context()
+	logger := utils.LoggerFromContext(ctx)
+
+	resourceID, err := utils.ResourceIDFromContext(ctx)
+	if err != nil {
+		return utils.TrackError(err)
+	}
+
+	operation, err := f.resourcesDBClient.Operations(resourceID.SubscriptionID).Get(ctx, resourceID.Name)
+	if err != nil {
+		return utils.TrackError(err)
+	}
+
+	// Validate the identity retrieving the operation result is the
+	// same identity that triggered the operation. Return 404 if not.
+	if !f.OperationIsVisible(request, operation) {
+		logger.Info("operation result not visible to requester")
+		writer.WriteHeader(http.StatusNotFound)
+		return nil
+	}
+
+	_, err = arm.WriteJSONResponse(writer, http.StatusOK, database.ToStatus(operation))
+	if err != nil {
+		return utils.TrackError(err)
+	}
+	return nil
+}
+
+func (f *Frontend) OperationResult(writer http.ResponseWriter, request *http.Request) error {
+	ctx := request.Context()
+
+	versionedInterface, err := VersionFromContext(ctx)
+	if err != nil {
+		return utils.TrackError(err)
+	}
+
+	resourceID, err := utils.ResourceIDFromContext(ctx)
+	if err != nil {
+		return utils.TrackError(err)
+	}
+
+	operation, err := f.resourcesDBClient.Operations(resourceID.SubscriptionID).Get(ctx, resourceID.Name)
+	if err != nil {
+		return utils.TrackError(err)
+	}
+
+	// Validate the identity retrieving the operation result is the
+	// same identity that triggered the operation. Return 404 if not.
+	if !f.OperationIsVisible(request, operation) {
+		return arm.NewResourceNotFoundError(resourceID)
+	}
+
+	// Handle non-terminal statuses and (maybe?) failure/cancellation.
+	//
+	// XXX ARM requirements for failed async operations get fuzzy here.
+	//
+	//     My best understanding, based on a Stack Overflow answer [1], is
+	//     returning an Azure-AsyncOperation header will cause ARM to poll
+	//     that endpoint first.
+	//
+	//     If ARM finds the operation in a "Failed" or "Canceled" state,
+	//     it will propagate details from the response's "error" property.
+	//
+	//     If ARM finds the operation in a "Succeeded" state, ONLY THEN is
+	//     this endpoint called (if a Location header was also returned).
+	//
+	//     So for the "Failed or Canceled" case we just give a generic
+	//     "Internal Server Error" response since, in theory, this case
+	//     should never be reached.
+	//
+	//     [1] https://stackoverflow.microsoft.com/a/318573/106707
+	//
+	switch operation.Status {
+	case arm.ProvisioningStateSucceeded:
+		// Handled below.
+	case arm.ProvisioningStateFailed, arm.ProvisioningStateCanceled:
+		return fmt.Errorf("invalid operation status: %s", operation.Status)
+	default:
+		// Operation is still in progress.
+		AddLocationHeader(writer, request, operation.OperationID)
+		writer.WriteHeader(http.StatusAccepted)
+		return nil
+	}
+
+	// The response henceforth should be exactly as though the operation
+	// succeeded synchronously.
+
+	var successStatusCode int
+
+	switch operation.Request {
+	case database.OperationRequestCreate:
+		successStatusCode = http.StatusCreated
+	case database.OperationRequestUpdate:
+		successStatusCode = http.StatusOK
+	case database.OperationRequestDelete:
+		writer.WriteHeader(http.StatusNoContent)
+		return nil
+	case database.OperationRequestRequestCredential:
+		successStatusCode = http.StatusOK
+	case database.OperationRequestRevokeCredentials:
+		writer.WriteHeader(http.StatusNoContent)
+		return nil
+	default:
+		return fmt.Errorf("unhandled request type: %s", operation.Request)
+	}
+
+	var responseBody []byte
+
+	switch {
+	case operation.InternalID.Kind() == cmv1.BreakGlassCredentialKind:
+		csBreakGlassCredential, err := f.clusterServiceClient.GetBreakGlassCredential(ctx, operation.InternalID)
+		if err != nil {
+			return utils.TrackError(err)
+		}
+
+		responseBody, err = versionedInterface.MarshalHCPOpenShiftClusterAdminCredential(ocm.ConvertCStoAdminCredential(csBreakGlassCredential))
+		if err != nil {
+			return utils.TrackError(err)
+		}
+
+	case armhelpers.ResourceTypeEqual(operation.ExternalID.ResourceType, api.ClusterResourceType):
+		resultingInternalCluster, err := f.getInternalClusterFromStorage(ctx, operation.ExternalID)
+		if err != nil {
+			return utils.TrackError(err)
+		}
+		responseBody, err = arm.MarshalJSON(versionedInterface.NewHCPOpenShiftCluster(resultingInternalCluster))
+		if err != nil {
+			return utils.TrackError(err)
+		}
+
+	case operation.ExternalID.ResourceType.String() == api.NodePoolResourceType.String():
+		resultingInternalNodePool, err := f.getInternalNodePoolFromStorage(ctx, operation.ExternalID)
+		if err != nil {
+			return utils.TrackError(err)
+		}
+		responseBody, err = arm.MarshalJSON(versionedInterface.NewHCPOpenShiftClusterNodePool(resultingInternalNodePool))
+		if err != nil {
+			return utils.TrackError(err)
+		}
+
+	case operation.ExternalID.ResourceType.String() == api.ExternalAuthResourceType.String():
+		resultingInternalExternalAuth, err := f.getInternalExternalAuthFromStorage(ctx, operation.ExternalID)
+		if err != nil {
+			return utils.TrackError(err)
+		}
+		responseBody, err = arm.MarshalJSON(versionedInterface.NewHCPOpenShiftClusterExternalAuth(resultingInternalExternalAuth))
+		if err != nil {
+			return utils.TrackError(err)
+		}
+
+	default:
+		return fmt.Errorf("unsupported operation reference: %s", operation.ExternalID)
+	}
+
+	_, err = arm.WriteJSONResponse(writer, successStatusCode, responseBody)
+	if err != nil {
+		return utils.TrackError(err)
+	}
+	return nil
+}
 
 // AddAsyncOperationHeader adds an "Azure-AsyncOperation" header to the ResponseWriter
 // with a URL of the operation status endpoint.
